requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-phone-field==1.7.0
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.2
ez-setup==0.9
gunicorn==19.9.0
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
xlwt==1.3.0
